requirements.txt
django
djangorestframework
python-dotenv
Django